Add caching of .cabal and .ghc directories
This will cache parts of the cabal directory structure between builds. This should speed build when we don't need to reinstall dependencies.

If the build fails, the cache is obliterated. This is a bit of a sledgehammer approach, and we should really just purge the cache when there is a dependency problem.